Update from payment_plan branch

+++ app/Http/Controllers/PropertyImagesController.php
@@ -15,46 +15,63 @@ use Illuminate\Http\File;
 class PropertyImagesController extends Controller
 {
 	/**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Property  $property
-     * @return \Illuminate\Http\Response
-     */
-    public function remove_images(Request $request)
-    {
-		$images = $request->remove_image;
-		$property = Property::find($request->prop);
-		
-		foreach($images as $image) {
-			$removeImage = PropertyImages::find($image);
-			
-			if($removeImage->delete()) {
-				Storage::delete($removeImage->path);
+	 * Remove the specified resource from storage.
+	 *
+	 * @param  \App\Property  $property
+	 * @return \Illuminate\Http\Response
+	 */
+	public function remove_images(Request $request)
+	{
+		if(isset($request->remove_image)) {
+			$images = $request->remove_image;
+
+			foreach($images as $image) {
+				$removeImage = PropertyImages::find($image);
+
+				if($removeImage->delete()) {
+					Storage::delete($removeImage->path);
+				}
 			}
 		}
 
-		return redirect()->action('PropertyController@index', $property)->with('status', 'Property Image(s) Deleted Successfully');
-    }
-	
+		if(isset($request->remove_video)) {
+			$videos = $request->remove_video;
+
+			foreach($videos as $video) {
+				$removeVideo = PropertyVideos::find($video);
+
+				if($removeVideo->delete()) {
+					Storage::delete($removeVideo->path);
+				}
+			}
+		}
+
+		return redirect()->back()->with('status', 'Property Media Items Deleted Successfully');
+	}
+
 	/**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Property  $property
-     * @return \Illuminate\Http\Response
-     */
-    public function remove_videos(Request $request)
-    {
-		$videos = $request->remove_video;
-		$property = Property::find($request->prop);
-		
-		foreach($videos as $video) {
-			$removeVideo = PropertyVideos::find($video);
-			
-			if($removeVideo->delete()) {
-				Storage::delete($removeVideo->path);
-			}
+	 * Remove the specified resource from storage.
+	 *
+	 * @param  \App\Property  $property
+	 * @return \Illuminate\Http\Response
+	 */
+	public function default_image(Request $request)
+	{
+		$image = PropertyImages::find($request->PropertyImages);
+		$property = $image->property;
+		$image->default_photo = 'Y';
+
+		if($property->medias()->withTrashed()->where('default_photo', 'Y')) {
+			// Make all default images null
+			$property->medias()->withTrashed()->where('default_photo', 'Y')->update([
+				'default_photo' => null
+			]);
+
+			if($image->save()) {}
+		} else {
+			if($image->save()) {}
 		}
 
-		return redirect()->action('PropertyController@index', $property)->with('status', 'Property Video(s) Deleted Successfully');
-    }
+		return $image;
+	}
 }
